+++ R/plot_boxplot.r
@@ -5,6 +5,7 @@
 #' @param by feature name to be broken down by. If selecting a continuous feature, boxplot will be grouped by 5 equal ranges, otherwise, all existing categories for a discrete feature.
 #' @param binary_as_factor treat binary as categorical? Default is \code{TRUE}.
 #' @param geom_boxplot_args a list of other arguments to \link{geom_boxplot}
+#' @param geom_jitter_args a list of other arguments to \link{geom_jitter}. If empty, \link{geom_jitter} will not be added.
 #' @param scale_y scale of original y axis (before \code{coord_flip}). See \link{scale_y_continuous} for all options. Default is \code{continuous}.
 #' @param title plot title
 #' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
@@ -27,10 +28,14 @@
 #' skew <- data.frame(y = rep(c("a", "b"), 500), replicate(4L, rbeta(1000, 1, 5000)))
 #' plot_boxplot(skew, by = "y", ncol = 2L)
 #' plot_boxplot(skew, by = "y", scale_y = "log10", ncol = 2L)
+#' 
+#' # Plot with `geom_jitter`
+#' plot_boxplot(iris, by = "Species", ncol = 2L, geom_jitter_args = list(width = NULL)) # Turn on with default settings
 
 plot_boxplot <- function(data, by,
                          binary_as_factor = TRUE,
                          geom_boxplot_args = list(),
+                         geom_jitter_args = list(),
                          scale_y = "continuous",
                          title = NULL,
                          ggtheme = theme_gray(), theme_config = list(),
@@ -62,11 +67,13 @@ plot_boxplot <- function(data, by,
     parallel = parallel,
     X = layout,
     FUN = function(x) {
-      ggplot(dt2[variable %in% feature_names[x]], aes(x = by_f, y = value)) +
+      base_plot <- ggplot(dt2[variable %in% feature_names[x]], aes(x = by_f, y = value)) +
         do.call("geom_boxplot", geom_boxplot_args) +
         do.call(paste0("scale_y_", scale_y), list()) +
         coord_flip() +
         xlab(by)
+      if (!identical(geom_jitter_args, list())) base_plot <- base_plot + do.call("geom_jitter", geom_jitter_args)
+      base_plot
     }
   )
   ## Plot objects